+++ packages/fetch-router/CHANGELOG.md
@@ -24,6 +24,7 @@ This is the changelog for [`fetch-router`](https://github.com/remix-run/remix/tr
   createRedirectResponse(location, status)
   ```
 
+- BREAKING CHANGE: Rename `InferRequestHandler` => `BuildRequestHandler`
 - Add `exclude` option to `resource()` and `resources()` route map helpers (#10858)
 
 ## v0.11.0 (2025-11-21)